task1_multi_agent/requirements.txt
streamlit
langchain
langchain-community
duckduckgo-search
python-dotenv
google-generativeai
langchain-google-genai
